attendance_system/attendance/face_recognition.py
# Placeholder for face recognition functionality
# This is a temporary implementation to allow the system to run without the face_recognition library

import base64
import os
import pickle
from io import BytesIO
from PIL import Image
import numpy as np
from django.conf import settings
from users.models import CustomUser

class FaceRecognitionSystem:
    def __init__(self):
        self.known_face_encodings = []
        self.known_face_names = []
        self.known_face_ids = []
        print("Face recognition system initialized (placeholder)")
    
    def load_encodings(self):
        """Load known face encodings from the database - placeholder implementation."""
        users = CustomUser.objects.filter(face_encoding__isnull=False)
        for user in users:
            self.known_face_names.append(user.get_full_name())
            self.known_face_ids.append(user.id)
            # In a real implementation, we would load the actual face encodings here
        
        print(f"Loaded {len(self.known_face_names)} user profiles with face encodings")
    
    def recognize_from_base64(self, base64_image):
        """
        Placeholder for face recognition from base64 encoded image.
        In this mock implementation, it will return sample students to simulate recognition.
        
        Args:
            base64_image (str): Base64 encoded image string
        
        Returns:
            list: List of recognized faces (mock data)
        """
        # Load users if not already loaded
        if not self.known_face_ids:
            self.load_encodings()
            
        # If we have no users, return empty
        if not self.known_face_ids:
            return []
            
        # For demonstration, return recognized users with face encodings 
        results = []
        users_with_encodings = CustomUser.objects.filter(face_encoding__isnull=False)
        
        for user in users_with_encodings[:3]:  # Limit to first 3 for performance
            results.append({
                'user_id': user.id,
                'name': user.get_full_name(),
                'confidence': 0.85
            })
            
        return results
    
    def recognize_face(self, frame):
        """
        Placeholder for face recognition in an image frame.
        
        Args:
            frame: Image as numpy array
        
        Returns:
            list: Mock recognition results
        """
        # Similar mock as above
        if not self.known_face_ids:
            self.load_encodings()
            
        if not self.known_face_ids:
            return []
            
        results = []
        users_with_encodings = CustomUser.objects.filter(face_encoding__isnull=False)
        
        for user in users_with_encodings[:2]:  # Limit to first 2 for performance
            results.append({
                'user_id': user.id,
                'name': user.get_full_name(),
                'confidence': 0.9
            })
            
        return results
    
    def add_face_encoding(self, user, image_base64):
        """
        Store a mock face encoding for the user.
        
        Args:
            user: CustomUser object
            image_base64: Base64 encoded image string
        
        Returns:
            tuple: (True, message) with explanation
        """
        try:
            # Remove the data URL prefix
            if ',' in image_base64:
                image_base64 = image_base64.split(',')[1]
                
            # In a real implementation, we would:
            # 1. Decode the image
            # 2. Detect face
            # 3. Generate face encoding using face_recognition or similar library
            # 4. Store the encoding
            
            # For now, just store a placeholder encoding to indicate this user has registered
            mock_encoding = pickle.dumps(np.random.rand(128))  # Typical face encoding is 128-dim vector
            
            # Save to user model
            user.face_encoding = mock_encoding
            user.save()
            
            return True, "Face registered successfully"
            
        except Exception as e:
            return False, f"Error registering face: {str(e)}"

# Singleton instance
face_recognition_system = FaceRecognitionSystem() 