attendance_system/requirements.txt
# Core Framework
Django==5.0.2
mysqlclient==2.2.0

# Image Processing
Pillow==10.0.1
opencv-python==4.8.0.76
dlib==19.24.2
face-recognition==1.3.0

# Data Processing
numpy==1.24.3
python-dateutil==2.8.2

# Excel Handling
XlsxWriter==3.1.0

# Development Tools
python-dotenv==1.0.0
django-debug-toolbar==4.2.0

# Security
django-cors-headers==4.3.1
django-allauth==0.57.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Additional Dependencies
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
requests==2.31.0
urllib3==2.1.0 